requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.28.3
PyJWT>=2.10.1,<3
python-dotenv==1.0.1
pydantic>=2.11.7,<3
asyncpg>=0.29.0,<1
boto3>=1.34.0
